Update FeatureSelectionAlgorithm.ipynb
Removed columns with standard deviation of zero (all values are the same). Introduced iterations of the finding the best merit algorithm. This way we can see features from most merit to least. May need to introduce randomness in picking the "best feature" so that we can find what groups of features provide the best merit.