New algorithm for extracting features from pcaps
added code to FeatureSelectionAlgorithm to get features from pcaps, decided to scrap that plan and coded new algorithm for this in GetFeaturesFromPcaps file. In and out folders holder input output for the original code in FeatureSelection but were too big to push to github so keeping all files except for a few examples on person computer.